fix remaining known bugs from 6.0->4.0 shim Bugfixes 1. Improved whitespace consumption in JsonParser 2. Added tests to make sure JsonParser throws errors 	on some bad inputs 3. Improved ordering of types in JsonSchema 4. Fixed bugs from 6.0->4.0 shim in JsonTabularize Minor changes 1. Parsing multiple files with the JsonGrepper is 	parallelized. 2. integer JNodes are converted to doubles before 	being compared to other integers or to doubles. 	This will avoid some weirdness related to order of 	operands. Additions 1. Reset button for Settings form to return settings 	to their defaults

+++ JsonToolsNppPlugin/JSONTools/JsonTabularize.cs
@@ -323,7 +323,7 @@ namespace JSON_Tools.JSON_Tools
 				{
 					if (node is int && (int)node == 0)
 					{
-						patharr.children.Add(new JNode(0, Dtype.FLOAT, 0));
+						patharr.children.Add(new JNode(0L, Dtype.INT, 0));
 					}
 					else
 					{
